requirements.txt
certifi==2020.6.20
chardet==3.0.4
httplib2==0.18.1
idna==2.10
nose==1.3.7
oauth2==1.9.0.post1
plurk-oauth==0.9.2
python-dotenv==0.14.0
requests==2.24.0
urllib3==1.25.10
